requirements.txt
drf-spectacular
djoser~=2.2.0
django~=4.2.1
djangorestframework~=3.14.0
cryptography~=40.0.2
pip~=21.3.1
wheel~=0.37.1
pytz~=2023.3
PyYAML~=6.0
attrs~=23.1.0
asgiref~=3.6.0
sqlparse~=0.4.4
requests~=2.30.0
certifi~=2023.5.7
oauthlib~=3.2.2
idna~=3.4
urllib3~=2.0.2
pycparser~=2.21
jsonschema~=4.17.3
pyrsistent~=0.19.3
setuptools~=60.2.0
PyJWT~=2.6.0
uritemplate~=4.1.1
inflection~=0.5.1
psycopg2